incubyteTest/requirements.txt
asgiref==3.5.0
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
coverage==6.3.2
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.3
django-heroku==0.3.1
djangorestframework==3.13.1
executing==0.8.3
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.1.1
jedi==0.18.1
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.28
psycopg2==2.9.3
pure-eval==0.2.2
py==1.11.0
Pygments==2.11.2
pyparsing==3.0.7
pytest==7.1.1
pytest-cov==3.0.0
pytest-django==4.5.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
whitenoise==6.0.0
